Combines contracts and updates them.
Game automatically restarts after Player 2 makes a move.

Event is printed showing the moves the two players made and the result.

Bets are calculated differently -> proportional to the amount you put
in.

Refund is given if no one wins the bet.

Fee is kept in bets, and that can be withdrawn by the contract owner.